+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.1.4] - 2022-03-21
 ### Changed
 - added freeCompilerArgs
+- ignore genesis-kotlin when applied to non java project
 
 ## [1.1.3] - 2022-03-18
 ### Changed