+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.5] - 2022-04-04
+### Changed
+- changed freeCompilerArgs to ListProperty
+- updated to kotlin 1.6.20
+- updated to gradle 7.4.2
+- updated gradle plugin com.gradle.plugin-publish to 0.21.0
+
 ## [1.1.4] - 2022-03-21
 ### Changed
 - added freeCompilerArgs